Code clean-upify. Add onFocus callback. CSS-Modulefy.
